requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-generativeai==0.7.2
flask-login==0.6.3
flask-wtf==1.2.1
python-dotenv==1.0.1
markdown==3.6
